Change Cookie notice
Rather than showing the notification by default, and checking if the cookie is set in localestorage, it is now hidden by default, and then checks, to see if the cookie is set. It will then change to show it, if not. This is done to prevent it first being rendered, and quickly disappearing again.